Create separation between reader and writer methods
