+++ lib/data/book.json
@@ -1,2 +1,1 @@
-["Sapiens","2012-10-31","News","Yuval","Harari","Online shop","Universal","white","John Hopkins","good"]
-["Lord of the Rings","2000-01-01","Fiction","Agatha","Christie","From a friend","United","black","Peter Sellers","bad"]
+["Sapiens","2000-01-01","News","Yuval","Harari","From a friend","Universal","white","Peter Sellers","good"]
