+++ lib/data/game.json
@@ -1,2 +1,1 @@
-["Super Mario","1980-10-02","Romance","Alexey","Pajitnov","Google Play","Nintendo","yellow","true","2022/09/10"]
-["Tetris","1974-08-16","Classic","Alexey","Pajitnov","Online shop","Nintendo","yellow","true","1999/12/31"]
+["Super Mario","1986-10-02","Comedy","Alexey","Pajitnov","Google Play","Nintendo","yellow","true","2022/09/10"]
