+++ lib/data/movie.json
@@ -1,2 +1,1 @@
-["Shawshank Redemption","1986-06-04","Fiction","Charlie","Chaplin","Amazon Prime Video","Warner","blue","false"]
-["Fried green tomatoes","1982-06-04","Fiction","Alexey","Pajitnov","Online shop","United","black","false"]
+["Shawshank Redemption","2022-02-22","Fiction","Charlie","Chaplin","Amazon Prime Video","Warner","blue","true"]
