+++ lib/data/music.json
@@ -1,2 +1,1 @@
-["Thriller","1982-10-25","Romance","Elvis","Presley","Apple iTunes","Sony","red","true"]
-["The Joshua tree","1986-11-14","Classic","Elvis","Presley","Apple iTunes","Universal","white","true"]
+["Thriller","2001-01-01","Romance","Elvis","Presley","Online shop","Sony","red","true"]
